🛡️ Sentinel: harden identifier generation and fix data leakage
- Implement centralized `generateSecureId` using `crypto.randomUUID()` in `src/lib/id-generator.ts`
- Refactor logger, analytics, and service utilities to use secure IDs
- Replace insecure `console.log` with `logger.debug` in session analytics
- Add unit tests for secure ID generation
- Verified with `pnpm run check`, `pnpm build:cloudflare`, and `pnpm security:check`
