+++ src/lib/analytics.ts
@@ -17,7 +17,7 @@
 
 import { createLogger } from '@/lib/logger';
 import { EnvLoader } from '@/lib/config/environment';
-import { generateSecureId } from '@/lib/utils';
+import { generateSecureId } from './id-generator';
 
 /**
  * Event categories for analytics
